Update: Added ip tracking, secondary login, menu builder, email notices
This update expands the configurability of the authentication system in
various ways, including allowing a second login credential, ip address
tracking, and some account policy enforcement systems. Included in this
update are some basic account observer driven email notifications.

Also, this update provided an initial configuration file based menu
builder system with security assessments, similar to the solution used
for the form builder system.

Finally, this update furthers the separation between the laravel
package and the enraiged system, as the intent is to eventually split
the Enraiged namespace into a distinct composer package.